requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
influxdb==5.0.0
python-dateutil==2.6.1
python-gitlab==1.2.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
